refactor(studio): split document repository search
