+++ src/contexts/studio/infrastructure/repository/document.py
@@ -5,7 +5,6 @@ from typing import TYPE_CHECKING
 from sqlalchemy.orm import selectinload
 
 from src.contexts.studio.infrastructure.repository.common import (
-    Any,
     Document,
     DocumentDto,
     DocumentRevision,
@@ -24,11 +23,14 @@ from src.contexts.studio.infrastructure.repository.common import (
     select,
     text,
 )
+from src.contexts.studio.infrastructure.repository.document_search import (
+    DocumentSearchRepositoryMixin,
+)
 
 __all__ = ["DocumentRepositoryMixin"]
 
 
-class DocumentRepositoryMixin:
+class DocumentRepositoryMixin(DocumentSearchRepositoryMixin):
     database: StudioDatabase
 
     if TYPE_CHECKING:
@@ -223,29 +225,6 @@ class DocumentRepositoryMixin:
             raise InvalidOperation("Document revision chain is invalid.")
         return revision
 
-    def _refresh_search(
-        self,
-        session: Session,
-        document: Document,
-        revision: DocumentRevision,
-    ) -> None:
-        session.execute(
-            text("DELETE FROM document_search WHERE document_id = :document_id"),
-            {"document_id": document.id},
-        )
-        session.execute(
-            text(
-                "INSERT INTO document_search(document_id, project_id, title, content) "
-                "VALUES (:document_id, :project_id, :title, :content)"
-            ),
-            {
-                "document_id": document.id,
-                "project_id": document.project_id,
-                "title": document.title,
-                "content": revision.content_markdown,
-            },
-        )
-
     def get_revision(
         self,
         project_id: str,
@@ -313,24 +292,3 @@ class DocumentRepositoryMixin:
             project.updated_at = now
             session.flush()
             return [_document_dto(documents[doc_id]) for doc_id in document_ids]
-
-    def search_documents(
-        self,
-        project_id: str,
-        query: str,
-        *,
-        owner_id: str | None,
-        guest_session_id: str | None,
-    ) -> list[dict[str, Any]]:
-        with self.database.session() as session:
-            self._project(session, project_id, owner_id, guest_session_id)
-            rows = session.execute(
-                text(
-                    "SELECT document_id, title, snippet(document_search, 3, '', "
-                    "'', ' … ', 16) AS excerpt "
-                    "FROM document_search WHERE project_id = :project_id "
-                    "AND document_search MATCH :query ORDER BY rank LIMIT 30"
-                ),
-                {"project_id": project_id, "query": query},
-            ).mappings()
-            return [dict(row) for row in rows]
